Improves Frequency computation.

+++ src/app/algo_counters.cpp
@@ -10,7 +10,7 @@
 #include <numeric>
 
 //-----------------------------------------------------------------------------
-double algo_counters::arithmeticSum(const CountersMap& values)
+int algo_counters::arithmeticSum(const CountersMap& values)
 {
   return std::accumulate(values.begin(), values.end(), 0);
 }
